PVD/requirements.txt
pillow>=8.2.0
loguru>=0.5.3
numpy>=1.19.5
